# sass/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MIS API 文档",
        "description": "伟明环保设备有限公司SIS系统 API 文档 api/documentation  php artisan l5-swagger:generate",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/auth/refresh": {
            "get": {
                "tags": [
                    "权限认证auth"
                ],
                "summary": "刷新token",
                "description": "使用说明：刷新token",
                "operationId": "refresh",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/alarm/page": {
            "get": {
                "tags": [
                    "报警alarm"
                ],
                "summary": "分页获取数据列表",
                "description": "使用说明：分页获取数据列表",
                "operationId": "alarm-index",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "num",
                        "in": "query",
                        "description": "每页数据量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "内容",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed"
                    }
                }
            }
        },
...
        "/api/alarm-grade": {
            "get": {
                "tags": [
                    "报警等级alarm-grade"
                ],
                "summary": "分页获取数据列表",
                "description": "使用说明：分页获取数据列表",
                "operationId": "alarm-grade-index",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "num",
                        "in": "query",
                        "description": "每页数据量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "关键字中文名搜索",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "错误代码，0：为没有错误",
                                            "type": "integer",
                                            "default": "0"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AlarmGrade"
                                            }
                                        },
                                        "message": {
                                            "description": "错误消息",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "报警等级alarm-grade"
                ],
                "summary": "新增单条数据",
                "description": "使用说明：新增单条数据",
                "operationId": "alarm-grade-store",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "中文名字",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "描述",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_value",
                        "in": "query",
                        "description": "最小值",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "max_value",
                        "in": "query",
                        "description": "最大值",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "store succeed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "错误代码，0：为没有错误",
                                            "type": "integer",
                                            "default": "0"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AlarmGrade"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "description": "错误消息",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/alarm-grade/{id}": {
            "get": {
                "tags": [
                    "报警等级alarm-grade"
                ],
                "summary": "获取详细信息",
                "description": "使用说明：获取详细信息",
                "operationId": "alarm-grade-show",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AlarmGrade主键",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "错误代码，0：为没有错误",
                                            "type": "integer",
                                            "default": "0"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AlarmGrade"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "description": "错误消息",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "报警等级alarm-grade"
                ],
                "summary": "修改",
                "description": "使用说明：修改单条数据",
                "operationId": "alarm-grade-update",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AlarmGrade主键",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "中文名字",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "描述",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_value",
                        "in": "query",
                        "description": "最小值",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "max_value",
                        "in": "query",
                        "description": "最大值",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update succeed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "错误代码，0：为没有错误",
                                            "type": "integer",
                                            "default": "0"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AlarmGrade"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "description": "错误消息",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "报警等级alarm-grade"
                ],
                "summary": "删除单条数据",
                "description": "使用说明：删除单条数据",
                "operationId": "alarm-grade-destroy",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AlarmGrade主键",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "detroy succeed"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AlarmGrade": {
                "title": "AlarmGrade model",
                "description": "AlarmGrade model",
                "properties": {
                    "id": {
                        "description": "创建数据模型",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "min_value": {
                        "type": "string"
                    },
                    "max_value": {
                        "type": "string"
                    },
                    "orgnization_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}